requirements.txt
agno
schedule
imapclient
ollama
aiosmtplib
psycopg2-binary
python-dotenv
